backend -- cote base termine

+++ src/main/java/com/analytique/gestion_analytique/database/repository/CentreRepo.java
@@ -2,11 +2,11 @@ package com.analytique.gestion_analytique.database.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.analytique.gestion_analytique.database.entity.Centre;
+import com.analytique.gestion_analytique.database.entity.centre.Centre;
 
 /**
  * CentreRepo
  */
-public interface CentreRepo extends JpaRepository<Centre,Long>{
+public interface CentreRepo extends JpaRepository<Centre,Integer>{
 	
 }
