+++ src/main/java/com/analytique/gestion_analytique/database/repository/NatureCentreRepo.java
@@ -2,7 +2,7 @@ package com.analytique.gestion_analytique.database.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.analytique.gestion_analytique.database.entity.NatureCentre;
+import com.analytique.gestion_analytique.database.entity.centre.NatureCentre;
 
-public interface NatureCentreRepo extends JpaRepository<NatureCentre,Long>{
+public interface NatureCentreRepo extends JpaRepository<NatureCentre,Integer>{
 }
