+++ src/main/java/com/analytique/gestion_analytique/database/repository/ProductionRepo.java
@@ -2,8 +2,8 @@ package com.analytique.gestion_analytique.database.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.analytique.gestion_analytique.database.entity.Production;
+import com.analytique.gestion_analytique.database.entity.production.Production;
 
-public interface ProductionRepo extends JpaRepository<Production,Long>{
+public interface ProductionRepo extends JpaRepository<Production,Integer>{
 
 }
